[circuit] Separate model, instance & subckt modules
