+++ tests/circuit/test_library.py
@@ -1,14 +1,14 @@
 import pytest
 
 from align.circuit.library import Library
-from align.circuit import device
+from align.circuit import model
 
 @pytest.fixture
 def library():
     return Library()
 
 def test_library_registration(library):
-    device.Model(
+    model.Model(
         name='TEST', prefix = 'M',
         pins=['D', 'G', 'S', 'B'],
         parameters={'W': 0, 'L': 0, 'NFIN': 1},
